+++ src/Netnr.ClassLibrary/Netnr.Sample/Startup.cs
@@ -1,0 +1,92 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
+using System.Linq;
+
+namespace Netnr.Sample
+{
+    public class Startup
+    {
+        public Startup(IConfiguration configuration, IHostEnvironment env)
+        {
+            GlobalTo.Configuration = configuration;
+            GlobalTo.HostEnvironment = env;
+        }
+
+        /// This method gets called by the runtime. Use this method to add services to the container.
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddControllersWithViews();
+
+            services.AddControllers().AddNewtonsoftJson(options =>
+            {
+                //Action原样输出JSON
+                options.SerializerSettings.ContractResolver = new Newtonsoft.Json.Serialization.DefaultContractResolver();
+                //日期格式化
+                options.SerializerSettings.DateFormatString = "yyyy-MM-dd HH:mm:ss";
+            });
+            
+            //配置swagger
+            services.AddSwaggerGen(c =>
+            {
+                c.SwaggerDoc("v1", new OpenApiInfo
+                {
+                    Title = "Netnr.Sample"
+                });
+
+                "Sample,Fast,Core".Split(',').ToList().ForEach(x =>
+                {
+                    c.IncludeXmlComments(System.AppContext.BaseDirectory + "Netnr." + x + ".xml", true);
+                });
+            });
+
+            //session
+            services.AddSession();
+        }
+
+        /// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        public void Configure(IApplicationBuilder app, IHostEnvironment env, IMemoryCache cache)
+        {
+            Core.CacheTo.memoryCache = cache;
+
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+            else
+            {
+                app.UseExceptionHandler("/Home/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                app.UseHsts();
+            }
+
+            //配置swagger
+            app.UseSwagger().UseSwaggerUI(c =>
+            {
+                c.DocumentTitle = "Netnr.Sample";
+                c.SwaggerEndpoint("/swagger/v1/swagger.json", "Netnr.Sample");
+            });
+
+            app.UseHttpsRedirection();
+            app.UseStaticFiles();
+
+            app.UseRouting();
+
+            app.UseAuthorization();
+
+            //session
+            app.UseSession();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
+                    name: "default",
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+            });
+        }
+    }
+}